+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to the Wazuh app project will be documented in this file.
 
 - Support for Wazuh 4.12.0
 - Add setting to limit the number of rows in csv reports [#7182](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7182)
-- Added a test to check the correctly handled of columns [#7268](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7268)
+- Added a test to check the tables columns fields are known (Only for development purposes) [#7268](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7268)
 
 ### Changed